feat: 未使用コードを消した。
